# pkg/swagger/api.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Chat-Server API",
    "version": "1.0.0",
    "contact": {
      "name": "mchekalov",
      "email": "mchekalov@gmail.com"
    }
  },
  "tags": [
    {
      "name": "ChatapiV1"
    }
  ],
  "host": "localhost:8080",
  "schemes": [
    "http",
    "https"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/chat-server/v1": {
      "delete": {
        "operationId": "ChatapiV1_Delete",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "type": "object",
              "properties": {}
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "required": false,
            "type": "string",
            "format": "int64"
          }
        ],
        "tags": [
          "ChatapiV1"
        ]
      }
    },
    "/chat-server/v1/create": {
      "post": {
        "operationId": "ChatapiV1_Create",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/chat_api_v1CreateResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/chat_api_v1CreateRequest"
            }
          }
        ],
        "tags": [
          "ChatapiV1"
        ]
      }
    },
    "/chat-server/v1/sendMessage": {
      "post": {
        "operationId": "ChatapiV1_SendMessage",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "type": "object",
              "properties": {}
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/chat_api_v1SendMessageRequest"
            }
          }
        ],
        "tags": [
          "ChatapiV1"
        ]
      }
    }
  },
  "definitions": {
    "chat_api_v1CreateRequest": {
      "type": "object",
      "properties": {
        "usernames": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "chatname": {
          "type": "string"
        }
      }
    },
    "chat_api_v1CreateResponse": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "format": "int64"
        }
      }
    },
    "chat_api_v1MessageWrap": {
      "type": "object",
      "properties": {
        "from": {
          "type": "string"
        },
        "text": {
          "type": "string"
        },
        "timestamp": {
          "type": "string",
          "format": "date-time"
        }
      }
    },
    "chat_api_v1SendMessageRequest": {
      "type": "object",
      "properties": {
        "info": {
          "$ref": "#/definitions/chat_api_v1MessageWrap"
        }
      }
    },
    "protobufAny": {
      "type": "object",
      "properties": {
        "@type": {
          "type": "string"
        }
      },
      "additionalProperties": {}
    },
    "rpcStatus": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    }
  }
}